refactor(ppba-driver,rp): type integer step parameters as usize
Cucumber's `{int}` parses through `FromStr`, so a step whose value indexes
a switch can declare `usize` directly instead of taking `i32` and casting
at every use. 37 steps elsewhere in the workspace already do this; these 46
did not, and paid for it with 48 `as usize` casts.

Retyping the parameter deletes each cast at its source rather than picking
a spelling for it — which is what makes this slice free of the wider
`as_conversions` policy question. It reaches the loop-bound cases too:
`for id in from..=to` and `for id in 0..count` yield `usize` once their
bounds do, so three `let id = id as usize;` lines go with them.

Behaviour change, and it is the one you want: a negative number in a
feature file now fails the step's own parse instead of wrapping to
18446744073709551615 and failing an ASCOM range check further in. Nothing
passes one today — the "invalid switch ID" scenarios use 16, 17, 100 and
999, and every negative literal in these feature files is a `{float}`.

Steps in these files that take an `i32` they never cast are left alone.

`session_steps.rs` keeps its single `as usize`: `expected_exposures` is a
`u32` local summed from the flat plan, not a step parameter, so it belongs
with the wider `as usize` question rather than here.

+++ services/ppba-driver/tests/bdd/steps/sensor_steps.rs
@@ -13,12 +13,8 @@ use cucumber::then;
 // ============================================================================
 
 #[then(expr = "switch {int} value should be approximately {float}")]
-async fn switch_value_approximately(world: &mut PpbaWorld, id: i32, expected: f64) {
-    let value = world
-        .switch_ref()
-        .get_switch_value(id as usize)
-        .await
-        .unwrap();
+async fn switch_value_approximately(world: &mut PpbaWorld, id: usize, expected: f64) {
+    let value = world.switch_ref().get_switch_value(id).await.unwrap();
     assert!(
         (value - expected).abs() < 0.1,
         "switch {id} value should be ~{expected}, got {value}"
@@ -26,12 +22,8 @@ async fn switch_value_approximately(world: &mut PpbaWorld, id: i32, expected: f6
 }
 
 #[then(expr = "switch {int} value should be in range {float} to {float}")]
-async fn switch_value_in_range(world: &mut PpbaWorld, id: i32, min: f64, max: f64) {
-    let value = world
-        .switch_ref()
-        .get_switch_value(id as usize)
-        .await
-        .unwrap();
+async fn switch_value_in_range(world: &mut PpbaWorld, id: usize, min: f64, max: f64) {
+    let value = world.switch_ref().get_switch_value(id).await.unwrap();
     assert!(
         (min..=max).contains(&value),
         "switch {id} value {value} should be in range {min}..={max}"
@@ -39,12 +31,8 @@ async fn switch_value_in_range(world: &mut PpbaWorld, id: i32, min: f64, max: f6
 }
 
 #[then(expr = "switch {int} value should be non-negative")]
-async fn switch_value_non_negative(world: &mut PpbaWorld, id: i32) {
-    let value = world
-        .switch_ref()
-        .get_switch_value(id as usize)
-        .await
-        .unwrap();
+async fn switch_value_non_negative(world: &mut PpbaWorld, id: usize) {
+    let value = world.switch_ref().get_switch_value(id).await.unwrap();
     assert!(
         value >= 0.0,
         "switch {id} value should be >= 0, got {value}"
@@ -52,12 +40,8 @@ async fn switch_value_non_negative(world: &mut PpbaWorld, id: i32) {
 }
 
 #[then(expr = "switch {int} value should be 0.0 or 1.0")]
-async fn switch_value_boolean_range(world: &mut PpbaWorld, id: i32) {
-    let value = world
-        .switch_ref()
-        .get_switch_value(id as usize)
-        .await
-        .unwrap();
+async fn switch_value_boolean_range(world: &mut PpbaWorld, id: usize) {
+    let value = world.switch_ref().get_switch_value(id).await.unwrap();
     assert!(
         value == 0.0 || value == 1.0,
         "switch {id} should be 0.0 or 1.0, got {value}"
@@ -65,12 +49,8 @@ async fn switch_value_boolean_range(world: &mut PpbaWorld, id: i32) {
 }
 
 #[then(expr = "switch {int} value should be positive")]
-async fn switch_value_positive(world: &mut PpbaWorld, id: i32) {
-    let value = world
-        .switch_ref()
-        .get_switch_value(id as usize)
-        .await
-        .unwrap();
+async fn switch_value_positive(world: &mut PpbaWorld, id: usize) {
+    let value = world.switch_ref().get_switch_value(id).await.unwrap();
     assert!(
         value > 0.0,
         "switch {id} value should be positive, got {value}"
